upd: refactoring fields and html properties

+++ week-08/reddit/src/main/java/com/greenfox/redditapp/repository/RedditRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RedditRepository extends CrudRepository<RedditItem, Float> {
+public interface RedditRepository extends CrudRepository<RedditItem, Long> {
 }
